ui-core: UCI: Move UCI logic into its own feature module
Creates a new feature module to group all UCI related components /
directives / services.
