+++ luci2-ui-core/src/src/app/shell/shell.module.ts
@@ -14,6 +14,7 @@ import { NavItemComponent } from './nav-menu/nav-item.component';
 import { NavMenuComponent } from './nav-menu/nav-menu.component';
 import { ShellComponent } from './shell/shell.component';
 import { MaterialModule } from 'app/material.module';
+import { UciModule } from 'app/uci/module';
 
 
 /**
@@ -30,6 +31,7 @@ import { MaterialModule } from 'app/material.module';
     BrowserAnimationsModule,
 
     MaterialModule,
+    UciModule
   ],
   declarations: [
     ShellComponent,
